Extract "near" assertion to own file
