+++ lib/referee.js
@@ -202,23 +202,7 @@ require("./assertions/is-array")(referee);
 require("./assertions/is-array-like")(referee);
 require("./assertions/keys")(referee);
 require("./assertions/exception")(referee);
-
-referee.add("near", {
-    assert: function (actual, expected, delta) {
-        return Math.abs(actual - expected) <= delta;
-    },
-    assertMessage: "${customMessage}Expected ${actual} to be equal to ${expected} +/- ${delta}",
-    refuteMessage: "${customMessage}Expected ${actual} not to be equal to ${expected} +/- ${delta}",
-    expectation: "toBeNear",
-    values: function (actual, expected, delta, message) {
-        return {
-            actual: actual,
-            expected: expected,
-            delta: delta,
-            customMessage: message
-        };
-    }
-});
+require("./assertions/near")(referee);
 
 referee.add("hasPrototype", {
     assert: function (actual, protoObj) {
